When optimizing hyperparameters, return the full set of parameters
There are two categories of hyperparameters:

- Hardcoded hyperparameters configured in the pipeline under `[:train-model
:hyperparameters]`.
- Hyperparameters search space for grid search or random search, configured in
`[:train-models :optimize :hyperparameter-search-space]`.

The old behavior of `optimize-hyperparameters` was to return only the selected
hyperparameters selected from `hyperparameter-search-space`. This caused an
issue in `curbside-prediction` as it assumed it was the full set of
hyperparameters, and then call `models/train` using only these, missing the
hardcoded one.

This PR change the return value of `optimize-hyperparameters` to be the whole
set of hyperparameters, as it is more natural and fixes the curbside-prediction issue.