Faster csv parsing (BREAKING CHANGES)
As described in https://github.com/RakutenReady/Team/issues/49909 parsing was
slow because for each row that is not a float, we throw and catch.

This PR changes this behavior by doing pre-validation before parsing a cell to a
number. This is done in a new function `parse-or-identity`, which differs
from the previous implementation regarding the following:
- If a string is not a number, the string is returned instead of throwing. This
does not change the behavior of `csv-to-maps` as the exception was handled at
this level
- BREAKING the ints are not longer parsed as doubles
- BREAKING the boolean are no longer parsed as `0.0` and `1.0`. This is legacy
code from old classification pipeline, which as never been used. Instead, we now
return boolean type values.

A quick benchmark on a production csv shows a speed-up of around 35%

```clojure
(comment
  (require '[criterium.core :as crit])
  (let [csv-path "/tmp/test-data.csv"]
    (println "new version:")
    (crit/quick-bench
     (do (csv-to-maps csv-path) nil))

    (let [old-parse-row-values
          (fn [csv-row]
            (map #(try
                    (parsing/parse-double %)
                    (catch Exception _e %))
                 csv-row))]
      (with-redefs [parse-row-values old-parse-row-values]
        (println "old version:")
        (crit/quick-bench
         (do (csv-to-maps csv-path) nil))))))
```

Output:
```
old version:
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 1.106209 sec
    Execution time std-deviation : 10.686383 ms
   Execution time lower quantile : 1.095409 sec ( 2.5%)
   Execution time upper quantile : 1.119044 sec (97.5%)
                   Overhead used : 2.242135 ns
new version:
Evaluation count : 6 in 6 samples of 1 calls.
             Execution time mean : 664.731381 ms
    Execution time std-deviation : 8.249613 ms
   Execution time lower quantile : 655.965658 ms ( 2.5%)
   Execution time upper quantile : 676.727322 ms (97.5%)
                   Overhead used : 2.242135 ns

Found 1 outliers in 6 samples (16.6667 %)
	low-severe	 1 (16.6667 %)
 Variance from outliers : 13.8889 % Variance is moderately inflated by outliers
```