Add the capability of training ranking xgboost models
https://github.com/RakutenReady/Team/issues/49826

This PR enables to load ranking datasets from csv files and to train models on
them. The format chosen for these datasets is the following:

- A single file contains both the features (X) and the labels (Y).
- Columns corresponding to labels are named `label-*`. All the other columns are considered as features.
- Each line corresponds to all the examples of a single member.

This is the responsibility of this library to create the one-hot encoding next
to the features before training the model. This allows to keep a low footprint
on disk.

See dummy-ranking.csv for an example.

This is yet to be integrated with the functions in the `curbside.ml.models` namespace.