Support cross validation and splitting for ranking datasets
https://github.com/RakutenReady/Team/issues/51847

For ranking, there is a group vector indicating how many successive examples are
part of the same group. For instance, for in-store recommendation, all examples
related to a same user are part of the same group.

To be able to train ranking models, we have to support splitting a dataset with
group information for k-fold cross validation and train-test splitting. Instead
of splitting across examples individually, we need to distributes groups in each
folds.

To do so, a new training-set namespace as been created. It is responsible of
loading, saving and splitting training sets. This is more maintainable that the
old code which was directly splitting lines of the csvs.

In the future, I would also like to do the following breaking changes:
- Rename `curbside.ml.training-sets.*` and
`curbside.ml.training-sets.training-set` to respectively
`curbside.ml.datasets.*` and `curbside.ml-datasets.dataset`. These contain
public functions used in the models manager and models serving, so it should be
changed in a separate release.
- Accept training-set maps in the public interface instead of CSVs. This would
semantically bind together the features maps, the weights and the groups, and
reduce the number of arguments to provide.