Compute metrics in metrics.clj instead of models.clj
https://github.com/RakutenReady/Team/issues/51929

Previously, most of the metrics computation was performed inside models.clj.
Model evaluation code computed the following map which was then passed to metrics.clj:

```
  {:abs-error ...
   :square-error ...
   :n ...
   :predictions ...}
```

Since this is specific to regression and classification, the interface with
metrics.clj has been changed to instead pass only the predictor-type,
predictions and labels. This will facilitate computing ranking metrics (in a
future PR).

This is a first PR including only this refactoring, to show that the unit tests
still work for all the predictor types. The next PR will actually add ranking
metric computation.