Rename training-set to dataset
While doing breaking changes to the model API for ranking models, it is a good
occasion to sneak in other breaking changes.

training-set is renamed to dataset. This is consistent with the new naming
convention used inside the pipeline since the introduction of the dataset pipelines.

Note that `curbside.ml.training-sets.*` namespaces are not used in models
serving.