Handle categories and missing values in Weka Instances conversion
The previous PR presented a dataset parser to create Weka Instances. This parser
did not handle properly missing values and categorical features. So, to handle
those, this PR goes back of using the csv-to-arff conversion. This is however
done in-memory, unlike how it was done before.

(Turns out the docstring of `csv-to-arff` was not true. Not all columns are
considered numerical, it handles categorical columns and missing values properly.)