Add full support for feature encoding training and infer
This adds the missing pieces for training and inferring using feature encoding.
This will be used to perform a one-hot encoding of the `offer-id` feature for
in-store.

The serialization/deserialisation of this encoding file is in edn. I
wanted to use JSON first but we cannot use integer keys in dictionaries.

For training, the feature encoding map is passed around inside the dataset. It
make sense to do so since having a dataset without its encoding is completely
useless.

BREAKING CHANGE the data/load-csv-files and dataset/save-csv-files were both
renamed, since they now save/load an edn file. This will impact the models
manager.

Thinking out loud, the only thing I do not like about this implementation is
that the feature encoding works differently than the feature scaling:
- The encoding is stored inside the dataset map, while the scaling factor is
passed around independently.
- The encoding consists of a single file all-in-one format, describing both the
configuration (which features to scale) and how to scale them. For scaling, we
have to pass around three things: the feature scaling config, the label scaling
config and a factor map.
- Feature encoding is done just in time, while scaling is performed ahead of
time in the pipeline.

I think that for all of the above, the feature scaling implementation got it
wrong. Also, from an external point of view, this proposed implementation for
the feature encoding will look consistent:
- In models-serving, we'll have to pass an additional keyword argument for
inference when using feature encoding. This is similar to scaling.
- In models-manager, passing the feature encoding inside the dataset will enable
to train without breaking the training interface.