Fix a regression for :classification models
This PR instrodcued a regression for classification models

https://github.com/RakutenReady/curbside-clojure-ml/commit/e07b551a08a05e35d8c2816a0bf3efb4a0f29a81#diff-24f612ff740c800af003707469dea3c7552852e311ae8c29ac1402f577fc8cf7L244

I thought that calling
```
(NominalPrediction. predicted (NominalPrediction/makeDistribution predicted 2)
```
was required in order to get a 0 or 1 value out of a classification model. Turns
out it isn't and the previous implementation in the linked PR was probably
broken.

Ideally, we would need to add tests in the clojure-ml library for the
classification models. However, since this is not used and very legacy, I think
ensuring that the tests pass in models-serving and models-manager is enough for now.

Also fix a docstring.