Handle unknown value when encoding features
It is possible that we get a query for recommending an unknown site for
in-store. We have to handle this case gracefully.

With these changes, we will return a `nil` prediction when the feature encoding
also returns `nil`. (it was throwing before, not the best...)

Note that if we wanted, we could still change the one-hot feature encoding to return a
default value (for instance a vector full of zeros), which is compatible with
this PR.

Also made public the infer-batch function which will be needed at serving time.