Switch from slf4j-simple to slf4j-log4j12
Fixes https://github.com/RakutenReady/Team/issues/53834

Even if the dependency is listed in the dev dependencies, it is still
imported by the consumers of the library.

We use log4j in both models manager and models serving. It is cleaner to
change the slf4j backend here than to exclude the slf4j-simple
dependency everywhere.