Relax the hyperparameter specs to accept int
The previous spec was to strict. It was refusing int values for
hyperparameters that can take floating point values. This is problematic
because values like `15.0` gets converted to the integer 15 for instance
after a round trip to JSON.

Note that most `double-in` specs where not specifying a min and a max
value, but only excluding NaN and Infinity. This is overkill because
these values are set manually in the configs.